ashmem_test: Cleanup and document fork tests
The tests that fork a child process and have the child process update
use ASSERT_EXIT() to fork and wait for the child process to terminate.
However, from looking at the code at a first pass, that is not entirely
obvious.

Therefore, document the fork tests and simplify some of the logic in
the tests.

No functional change.

Bug: 111903542
Change-Id: Iac923ccad9cea0fc7d823c6e2c46c47fbf6c0282
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>