ueventd: allow FirmwareHandler to run in a single process after coldboot
FirmwareHandler::HandleUevent internally parallelizes the core work
independently from the main ueventd process by fork. Thus, we have two
parallelization on going when we do parallelization in ueventd.

On the other hand, we are working on giving the ueventd an option to
parallelize its main loop with multiple threads. In that case, we don't
want the internal parallelization of FirmwareHandler since the main loop
runs in parallel in the first place. In addition, fork syscall does not
work well with multi-threaded main loop.

Thus, This CL makes it possible to disable the FirmwareHandler internal
parallelization. For the backward compatibility sake, this is disabled
by default, and even with it enabled, FirmwareHandler becomes a single
process only after the coldboot gets done.

Bug: 400592897
Test: Cuttlefish boots
Change-Id: I4b395c48ea3e696cd221a3033b0e6b8dc4afd62d