fs_mgr: enable quota by default
It was reported there is a security issue in Pixel device, it can be
hit by DoS attack by a malicious app trying to make use of all inodes
available in the system.

The root cause is filesystem hasn't fully enabled disk quota feature,
as by default, it only enabled accounting usage of all files inside
internal quota database sysfile, however, it missed to enable quota
limitation.

There are two ways to fix this:
1. fix in userspace
 - add "usrquota,grpquota,prjquota" mount option during mount
2. fix in kernel
 - pass both DQUOT_USAGE_ENABLED and DQUOT_LIMITS_ENABLED flag into
   f2fs_quota_enable()

To keep flexibility on kernel functionality, it's more proper to fix
this in userpspace.

Bug: 416567751
Test: make
Flag: EXEMPT bugfix
Change-Id: I00c56f80dab13587eb2fbb9755423af97c660e5f
Signed-off-by: Chao Yu <chaseyu@google.com>