prefetch: Handle large folios correctly
Newer kernel supports large folios in filesystems, where page cache
entries are added in the unit of (1 << order) pages. The initial support
was added in [1], where the size (order) was exposed in tracepoint.
Recently, EROFS started supporting large folios on compressed pages in
v6.11, which along with the v6.12 uprev to cause obviously missed pages
and large increase in pack size due to fragmentation.

Add support for parsing the order field, and reflect that in the page
records.

[1]: https://lore.kernel.org/all/20220116121822.1727633-4-willy@infradead.org/

Bug: 409692272
Test: cargo test
Test: prefetch.pack is 200KB and reads 724MB, vs. 1MB and read 400MB on ToT
Change-Id: I5f94f3bc8563f43875394cff3ce58916938e832b