Strengthen memory ordering for RefBase tag/untag.
We had one case of this:

    RBSAN: Can't unset flag 65536 when it's not set: 0

We may have the case where 'tag' is called on one
thread, but 'untag' is called on another thread.
If the memory is not updated in the CPU cache,
then the 'untag' may abort. Go for the stronger
semantics here since this is a resource acquisition,
and this clarity could make some error messages more
clear.

Test: libutils_binder_test
Fixes: 442050744
Flag: EXEMPT BUG_FIX
Change-Id: I0d26781f45c25ea8c8a50158230825ede2f98fbb