Terminate zygote_next on `ctl.stop`
Native Zygote (go/native-zygote) is a rewrite of Zygote primarily aimed
for supporting java-less processes, and is currently built behind a
build flag. The `zygote_next` process is the process for Native Zygote
and currently it is not stopped upon `ctl.stop`, which is inconsistent
with the behavior of the managed zygote process.

This causes some issues when doing something like `adb shell stop; adb
shell setprop libc.debug.malloc.options backtrace_enable_on_signal; adb
shell start` - the zygote_next process does not reflect the libc
property update since it is not restarted.

This commit fixes this issue by adding zygote_next to the list of
default services that are stopped. The addition is put behind an #ifdef
just like it is done in frameworks/base/core/jni/AndroidRuntime.cpp.

Test: zygote_next process is restarted upon stop and start command
Bug: 447306479
Flag: build.RELEASE_NATIVE_FRAMEWORK_PROTOTYPE
Change-Id: Ibfe1948376976c7b100089ec344a1c94642c4b7e