init disable reboot bootloader on pending OTA
When there is a pending OTA, we want to disable to reboot bootloader path when init panics. This is because OTA has a built in fallback mechanism. The extra reboot bootloader on userdebug builds will invalidate pstore logs and thus we'll lose the ability to diagnose rollbacks. In order to capture early failures in first stage init, we will move back installing crash handlers to after /metadata is mounted in order to check whether OTA is pending.

Bug: 394172800
Test: rollback ota on rango-userdebug build
Change-Id: I02714de5d0d60d7f5f581b2c0caed26d60635d08