healthd: automated reformatting
This change reformats the sources in the healthd directory:

    clang-format -i healthd/*.cpp healthd/*.h

These files contained a variety of incorrect formatting which makes it
more difficult to refactor because refactoring incorrectly-formatted
parts tends to demand fixing the formatting of nearby code which
increases diff noise. Make all of the formatting changes at once to
avoid that noise.

Test: m libbatterymonitor
Change-Id: Ib0f3038f829daf1b346c85c22e9cbb399b5dc465