healthd: BatteryMonitor: retain battery level on error
If there is an error reading the battery level, this change avoids
immediately acting as if the battery is empty and instead returns the
last-seen valid battery level as long as it was seen recently enough.
If seen too long ago or no valid value has even been seen, 0 is still
returned because the API offers no way to specify "don't know" for the
battery level.

Bug: 410996664
Test: on brya, force battery communications failure in EC firmware;
      when ENODEV is returned from the next battery level read, the
      new log message is printed and battery level is not reported as
      0 until some time is elapsed or the battery reports as absent.
Change-Id: If91f8b8f5cd0a766ff3a6d58b2845f799d298c99